session3_code/pyproject.toml
[project]
name = "albertschool-llmops"
version = "0.1.0"
description = "Course materials for the LLMOps class for Albert School!"
authors = ["Mathieu Soul <mathieus@sicara.com>"]
readme = "README.md"
requires-python = "==3.11.6"
dependencies = [
    "google-cloud-aiplatform[prediction]>=1.38.0",
    "google-cloud-storage>=2.10.0",
    "pandas>=2.3.2",
    "kfp>=2.0.0",
    "datasets>=4.1.1",
    "ipykernel>=6.30.1",
    "ragas>=0.3.5",
    "rouge-score>=0.1.2",
    "gcsfs>=2025.9.0",
    "tensorboard>=2.20.0",
    "rapidfuzz>=3.14.1",
    "sacrebleu>=2.5.1",
    "transformers==4.46.*",
    "accelerate>=1.10.1",
    "trl>=0.17.0",
    "peft==0.13.2",
]
